# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TradeInput {
  pair: string;
  direction: string;
  lotSize: number;
  slPips: number;
  /** @nullable */
  tpPips?: number | null;
  riskUSD: number;
  riskNGN: number;
  usdRate: number;
  /** @nullable */
  rr?: string | null;
  notes?: string;
}