# lib/api-zod/src/generated/types/tradeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TradeUpdate {
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  pnlUSD?: number | null;
  /** @nullable */
  pnlNGN?: number | null;
  notes?: string;
}